+++ CHANGELOG.md
@@ -8,9 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added 
-- Caculate the distance between LineSegment3D/LS3D, Arc3D/LS3D, Arc3D/Arc3D and between CylindricalFace3D too.
+- Calculate the distance between LineSegment3D/LS3D, Arc3D/LS3D, Arc3D/Arc3D and between CylindricalFace3D too.
 - Use PlaneFace3D with contours2D in a classic way and use it with contours3D with a 'from_contours3d' as CylindricalFace3D does.
+- Calculate the distance between CylindricalFace3D and PlaneFace3D.
+- Calculate the distance between CylindricalFace3D, PlaneFace3D and ToroidalFace3D.
+- contours2d.tessel_points which gives all points of a contour2d, and .points the end points of primitives.
+- Implementation of ConicalFace3D in Core and RevolvedProfile.
+- Implementation of SphericalFace3D in Core.
 
+### Changed
+- cut_contours in Face3D which take all points from a Contour2D, not one side like before. Furthermore, it is light and quick.
 
 ## [v0.1.10]
 - typings